@@ -7,9 +7,10 @@ module ShopifyApp
     included do
       include ShopifyApp::Localization
       include ShopifyApp::LoginProtection
+      include ShopifyApp::CsrfProtection
       include ShopifyApp::EmbeddedApp
-      before_action :login_again_if_different_shop
-      around_action :shopify_session
+      before_action :login_again_if_different_user_or_shop
+      around_action :activate_shopify_session
     end
   end
 end
