@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'rails/generators/base'
 
 module ShopifyApp
@@ -6,7 +7,7 @@ module ShopifyApp
       source_root File.expand_path('../templates', __FILE__)
 
       def generate_app_extension
-        template "marketing_activities_controller.rb", "app/controllers/marketing_activities_controller.rb"
+        template("marketing_activities_controller.rb", "app/controllers/marketing_activities_controller.rb")
         generate_routes
       end
 
