@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'rails/generators/base'
 
 module ShopifyApp
@@ -6,11 +7,11 @@ module ShopifyApp
       source_root File.expand_path('../templates', __FILE__)
 
       def create_app_proxy_controller
-        template 'app_proxy_controller.rb', 'app/controllers/app_proxy_controller.rb'
+        template('app_proxy_controller.rb', 'app/controllers/app_proxy_controller.rb')
       end
 
       def create_app_proxy_index_view
-        copy_file 'index.html.erb', 'app/views/app_proxy/index.html.erb'
+        copy_file('index.html.erb', 'app/views/app_proxy/index.html.erb')
       end
 
       def add_app_proxy_route
@@ -18,7 +19,7 @@ module ShopifyApp
           'config/routes.rb',
           File.read(File.expand_path(find_in_source_paths('app_proxy_route.rb'))),
           after: "mount ShopifyApp::Engine, at: '/'\n"
-        )  			 
+        )
       end
     end
   end
