@@ -1,8 +1,8 @@
+# frozen_string_literal: true
 class AppProxyController < ApplicationController
-   include ShopifyApp::AppProxyVerification
+  include ShopifyApp::AppProxyVerification
 
   def index
-    render layout: false, content_type: 'application/liquid'
+    render(layout: false, content_type: 'application/liquid')
   end
-
 end
