@@ -1,10 +1,11 @@
+# frozen_string_literal: true
 
-  namespace :app_proxy do
-    root action: 'index'
-    # simple routes without a specified controller will go to AppProxyController
-    
-    # more complex routes will go to controllers in the AppProxy namespace
-    # 	resources :reviews
-    # GET /app_proxy/reviews will now be routed to
-    # AppProxy::ReviewsController#index, for example
-  end
+namespace :app_proxy do
+  root action: 'index'
+  # simple routes without a specified controller will go to AppProxyController
+
+  # more complex routes will go to controllers in the AppProxy namespace
+  #   resources :reviews
+  # GET /app_proxy/reviews will now be routed to
+  # AppProxy::ReviewsController#index, for example
+end
