@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'rails/generators/base'
 
 module ShopifyApp
@@ -11,16 +12,12 @@ module ShopifyApp
       class_option :embedded, type: :string, default: 'true'
       class_option :api_version, type: :string, default: nil
 
-      def add_dotenv_gem
-        gem('dotenv-rails', group: [:test, :development])
-      end
-
       def create_shopify_app_initializer
         @application_name = format_array_argument(options['application_name'])
         @scope = format_array_argument(options['scope'])
         @api_version = options['api_version'] || ShopifyAPI::Meta.admin_versions.find(&:latest_supported).handle
 
-        template 'shopify_app.rb', 'config/initializers/shopify_app.rb'
+        template('shopify_app.rb', 'config/initializers/shopify_app.rb')
       end
 
       def create_session_store_initializer
@@ -28,28 +25,28 @@ module ShopifyApp
       end
 
       def create_and_inject_into_omniauth_initializer
-        unless File.exist? "config/initializers/omniauth.rb"
-          copy_file 'omniauth.rb', 'config/initializers/omniauth.rb'
+        unless File.exist?("config/initializers/omniauth.rb")
+          copy_file('omniauth.rb', 'config/initializers/omniauth.rb')
         end
 
         inject_into_file(
           'config/initializers/omniauth.rb',
           File.read(File.expand_path(find_in_source_paths('shopify_provider.rb'))),
-          after: "Rails.application.config.middleware.use OmniAuth::Builder do\n"
+          after: "Rails.application.config.middleware.use(OmniAuth::Builder) do\n"
         )
       end
 
       def create_embedded_app_layout
         return unless embedded_app?
 
-        copy_file 'embedded_app.html.erb', 'app/views/layouts/embedded_app.html.erb'
-        copy_file '_flash_messages.html.erb', 'app/views/layouts/_flash_messages.html.erb'
+        copy_file('embedded_app.html.erb', 'app/views/layouts/embedded_app.html.erb')
+        copy_file('_flash_messages.html.erb', 'app/views/layouts/_flash_messages.html.erb')
 
         if ShopifyApp.use_webpacker?
           copy_file('shopify_app.js', 'app/javascript/shopify_app/shopify_app.js')
           copy_file('flash_messages.js', 'app/javascript/shopify_app/flash_messages.js')
           copy_file('shopify_app_index.js', 'app/javascript/shopify_app/index.js')
-          append_to_file('app/javascript/packs/application.js', 'require("shopify_app")')
+          append_to_file('app/javascript/packs/application.js', "require(\"shopify_app\")\n")
         else
           copy_file('shopify_app.js', 'app/assets/javascripts/shopify_app.js')
           copy_file('flash_messages.js', 'app/assets/javascripts/flash_messages.js')
@@ -57,11 +54,11 @@ module ShopifyApp
       end
 
       def create_user_agent_initializer
-        template 'user_agent.rb', 'config/initializers/user_agent.rb'
+        template('user_agent.rb', 'config/initializers/user_agent.rb')
       end
 
       def mount_engine
-        route "mount ShopifyApp::Engine, at: '/'"
+        route("mount ShopifyApp::Engine, at: '/'")
       end
 
       def insert_hosts_into_development_config
