@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'test_helper'
 
 module Shopify
@@ -8,10 +9,9 @@ end
 
 module ShopifyApp
   class SessionsControllerTest < ActionController::TestCase
-
     setup do
       @routes = ShopifyApp::Engine.routes
-      ShopifyApp::SessionRepository.storage = ShopifyApp::InMemorySessionStore
+      ShopifyApp::SessionRepository.shop_storage = ShopifyApp::InMemoryShopSessionStore
       ShopifyApp.configuration = nil
       ShopifyApp.configuration.embedded_app = true
 
@@ -19,7 +19,8 @@ module ShopifyApp
 
       session['shopify.granted_storage_access'] = true
 
-      request.env['HTTP_USER_AGENT'] = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/69.0.3497.100 Safari/537.36'
+      request.env['HTTP_USER_AGENT'] = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_6) AppleWebKit/537.36 (KHTML,
+                 like Gecko) Chrome/69.0.3497.100 Safari/537.36'
     end
 
     test '#new redirects to the enable_cookies page if we can\'t set cookies and the user agent supports cookie partitioning' do
@@ -42,19 +43,30 @@ module ShopifyApp
     end
 
     test '#new renders the redirect layout if user agent is Shopify Mobile (Android)' do
-      request.env['HTTP_USER_AGENT'] = 'Shopify Mobile/Android/8.12.0 (Build 12005 with API 28 on Google Android SDK built for x86) MobileMiddlewareSupported Mozilla/5.0 (Linux; Android 9; Android SDK built for x86 Build/PSR1.180720.075; wv)  AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Chrome/69.0.3497.100 Mobile Safari/537.36'
+      request.env['HTTP_USER_AGENT'] = 'Shopify Mobile/Android/8.12.0 (Build 12005 with API 28 on Google
+                                        Android SDK built for x86) MobileMiddlewareSupported Mozilla/5.0
+                                        (Linux; Android 9; Android SDK built for x86 Build/PSR1.180720.075; wv)
+                                        AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Chrome/69.0.3497.100
+                                        Mobile Safari/537.36'
       get :new, params: { shop: 'my-shop' }
       assert_template 'shared/redirect'
     end
 
     test '#new renders the redirect layout if user agent is Shopify Mobile (iOS)' do
-      request.env['HTTP_USER_AGENT'] = 'Mozilla/5.0 (iPhone; CPU iPhone OS 12_0 like Mac OS X) AppleWebKit/ 604.1.21 (KHTML, like Gecko) Version/ 12.0 Mobile/17A6278a Safari/602.1.26 MobileMiddlewareSupported Shopify Mobile/iOS/8.12.0 (iPad4,7/com.shopify.ShopifyInternal/12.0.0)'
+      request.env['HTTP_USER_AGENT'] = 'Mozilla/5.0 (iPhone; CPU iPhone OS 12_0 like Mac OS X)
+                                        AppleWebKit/ 604.1.21 (KHTML, like Gecko) Version/ 12.0 Mobile/17A6278a
+                                        Safari/602.1.26 MobileMiddlewareSupported
+                                        Shopify Mobile/iOS/8.12.0 (iPad4,7/com.shopify.ShopifyInternal/12.0.0)'
       get :new, params: { shop: 'my-shop' }
       assert_template 'shared/redirect'
     end
 
     test '#new renders the redirect layout if user agent is Shopify POS (Android)' do
-      request.env['HTTP_USER_AGENT'] = 'com.jadedpixel.pos Shopify POS Dalvik/2.1.0 (Linux; U; Android 7.0; Android SDK built for x86 Build/NYC) POS - Debug 2.4.8 (f1d442c789)/3405 Mozilla/5.0 (Linux; Android 7.0; Android SDK built for x86 Build/NYC; wv) AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Chrome/64.0.3282.137 Safari/537.36'
+      request.env['HTTP_USER_AGENT'] = 'com.jadedpixel.pos Shopify POS Dalvik/2.1.0
+                  (Linux; U; Android 7.0; Android SDK built for x86 Build/NYC)
+                  POS - Debug 2.4.8 (f1d442c789)/3405 Mozilla/5.0
+                  (Linux; Android 7.0; Android SDK built for x86 Build/NYC; wv)
+                  AppleWebKit/537.36 (KHTML, like Gecko)Version/4.0 Chrome/64.0.3282.137 Safari/537.36'
       get :new, params: { shop: 'my-shop' }
       assert_template 'shared/redirect'
     end
@@ -69,6 +81,16 @@ module ShopifyApp
       shopify_domain = 'my-shop.myshopify.com'
       get :new, params: { shop: 'my-shop' }
       assert_redirected_to_top_level(shopify_domain)
+    end
+
+    test '#new stores root path when return_to url is absolute' do
+      get :new, params: { shop: 'my-shop', return_to: '//example.com' }
+      assert_equal '/', session[:return_to]
+    end
+
+    test '#new stores only relative path for return_to in the session' do
+      get :new, params: { shop: 'my-shop', return_to: '/page' }
+      assert_equal '/page', session[:return_to]
     end
 
     test '#new sets the top_level_oauth cookie if a valid shop param exists and user agent supports cookie partitioning' do
@@ -117,17 +139,18 @@ module ShopifyApp
     test "#new should render a full-page if the shop param doesn't exist" do
       get :new
       assert_response :ok
-      assert_match %r{Shopify App — Installation}, response.body
+      assert_match(/Shopify App — Installation/, response.body)
     end
 
     test "#new should render a full-page if the shop param value is not a shop" do
       non_shop_address = "example.com"
       get :new, params: { shop: non_shop_address }
       assert_response :ok
-      assert_match %r{Shopify App — Installation}, response.body
+      assert_match(/Shopify App — Installation/, response.body)
     end
 
-    ['my-shop', 'my-shop.myshopify.com', 'https://my-shop.myshopify.com', 'http://my-shop.myshopify.com'].each do |good_url|
+    ['my-shop', 'my-shop.myshopify.com', 'https://my-shop.myshopify.com',
+     'http://my-shop.myshopify.com'].each do |good_url|
       test "#create should authenticate the shop for the URL (#{good_url})" do
         shopify_domain = 'my-shop.myshopify.com'
         post :create, params: { shop: good_url }
@@ -135,7 +158,8 @@ module ShopifyApp
       end
     end
 
-    ['my-shop', 'my-shop.myshopify.io', 'https://my-shop.myshopify.io', 'http://my-shop.myshopify.io'].each do |good_url|
+    ['my-shop', 'my-shop.myshopify.io', 'https://my-shop.myshopify.io',
+     'http://my-shop.myshopify.io'].each do |good_url|
       test "#create should authenticate the shop for the URL (#{good_url}) with custom myshopify_domain" do
         ShopifyApp.configuration.myshopify_domain = 'myshopify.io'
         shopify_domain = 'my-shop.myshopify.io'
@@ -144,13 +168,22 @@ module ShopifyApp
       end
     end
 
-    ['myshop.com', 'myshopify.com', 'shopify.com', 'two words', 'store.myshopify.com.evil.com', '/foo/bar'].each do |bad_url|
+    ['myshop.com', 'myshopify.com', 'shopify.com', 'two words',
+     'store.myshopify.com.evil.com', '/foo/bar'].each do |bad_url|
       test "#create should return an error for a non-myshopify URL (#{bad_url})" do
         post :create, params: { shop: bad_url }
         assert_response :redirect
         assert_redirected_to '/'
         assert_equal I18n.t('invalid_shop_url'), flash[:error]
       end
+    end
+
+    test '#create should return an error for a non-myshopify URL when using JWT authentication' do
+      ShopifyApp.configuration.allow_jwt_authentication = true
+      post :create, params: { shop: 'invalid domain' }
+      assert_response :redirect
+      assert_redirected_to '/'
+      assert_equal I18n.t('invalid_shop_url'), flash[:error]
     end
 
     test "#create should render the login page if the shop param doesn't exist" do
@@ -229,7 +262,7 @@ module ShopifyApp
     def assert_redirected_to_top_level(shop_domain, expected_url = nil)
       expected_url ||= "/login?shop=#{shop_domain}\\u0026top_level=true"
 
-      assert_template 'shared/redirect'
+      assert_template('shared/redirect')
       assert_select '[id=redirection-target]', 1 do |elements|
         assert_equal "{\"myshopifyUrl\":\"https://#{shop_domain}\",\"url\":\"#{expected_url}\"}",
           elements.first['data-target']
