@@ -1,7 +1,6 @@
+# frozen_string_literal: true
 class HomeController < ApplicationController
-
   def index
     'index'
   end
-
 end
