@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'test_helper'
 require 'generators/shopify_app/controllers/controllers_generator'
 
@@ -12,5 +13,4 @@ class ControllersGeneratorTest < Rails::Generators::TestCase
     assert_file "app/controllers/shopify_app/sessions_controller.rb"
     assert_file "app/controllers/shopify_app/authenticated_controller.rb"
   end
-
 end
