@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'test_helper'
 require 'generators/shopify_app/shop_model/shop_model_generator'
 
@@ -14,7 +15,7 @@ class ShopModelGeneratorTest < Rails::Generators::TestCase
     run_generator
     assert_file "app/models/shop.rb" do |shop|
       assert_match "class Shop < ActiveRecord::Base", shop
-      assert_match "include ShopifyApp::SessionStorage", shop
+      assert_match "include ShopifyApp::ShopSessionStorage", shop
       assert_match(/def api_version\n\s*ShopifyApp\.configuration\.api_version\n\s*end/, shop)
     end
   end
@@ -29,7 +30,7 @@ class ShopModelGeneratorTest < Rails::Generators::TestCase
   test "updates the shopify_app initializer" do
     run_generator
     assert_file "config/initializers/shopify_app.rb" do |file|
-      assert_match "config.session_repository = Shop", file
+      assert_match "config.shop_session_repository = 'Shop'", file
     end
   end
 
@@ -39,5 +40,4 @@ class ShopModelGeneratorTest < Rails::Generators::TestCase
       assert_match "regular_shop:", file
     end
   end
-
 end
