@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'test_helper'
 
 class AppProxyVerificationController < ActionController::Base
@@ -7,7 +8,7 @@ class AppProxyVerificationController < ActionController::Base
   include ShopifyApp::AppProxyVerification
 
   def basic
-    head :ok
+    head(:ok)
   end
 end
 
@@ -21,18 +22,26 @@ class AppProxyVerificationTest < ActionController::TestCase
   end
 
   test 'no_signature' do
-    assert_not query_string_valid? 'shop=some-random-store.myshopify.com&path_prefix=%2Fapps%2Fmy-app&timestamp=1466106083'
+    assert_not query_string_valid?('shop=some-random-store.myshopify.com&'\
+                                   'path_prefix=%2Fapps%2Fmy-app&timestamp=1466106083')
   end
 
   test 'basic_query_string' do
-    assert query_string_valid? 'shop=some-random-store.myshopify.com&path_prefix=%2Fapps%2Fmy-app&timestamp=1466106083&signature=f5cd7233558b1c50102a6f33c0b63ad1e1072a2fc126cb58d4500f75223cefcd'
-    assert_not query_string_valid? 'shop=some-random-store.myshopify.com&path_prefix=%2Fapps%2Fmy-app&timestamp=1466106083&evil=1&signature=f5cd7233558b1c50102a6f33c0b63ad1e1072a2fc126cb58d4500f75223cefcd'
-    assert_not query_string_valid? 'shop=some-random-store.myshopify.com&path_prefix=%2Fapps%2Fmy-app&timestamp=1466106083&evil=1&signature=wrongwrong8b1c50102a6f33c0b63ad1e1072a2fc126cb58d4500f75223cefcd'
+    assert query_string_valid?('shop=some-random-store.myshopify.com&path_prefix=%2Fapps%2Fmy-app&timestamp='\
+      '1466106083&signature=f5cd7233558b1c50102a6f33c0b63ad1e1072a2fc126cb58d4500f75223cefcd')
+    assert_not query_string_valid?('shop=some-random-store.myshopify.com&path_prefix=%2Fapps%2Fmy-app&timestamp='\
+      '1466106083&evil=1&signature=f5cd7233558b1c50102a6f33c0b63ad1e1072a2fc126cb58d4500f75223cefcd')
+    assert_not query_string_valid?('shop=some-random-store.myshopify.com&path_prefix=%2Fapps%2Fmy-'\
+      'app&timestamp=1466106083&evil=1&signature=wrongwrong8b1c50102a6f33c0b63ad1e1072a2fc126cb58d4500f75223cefcd')
   end
 
   test 'query_string_complex_args' do
-    assert query_string_valid? 'shop=some-random-store.myshopify.com&path_prefix=%2Fapps%2Fmy-app&timestamp=1466106083&signature=bbf3aa60e098f08919a2ea4c64a388414f164e6a117a63b03479ac7aa9464b4f&foo=bar&baz[1]&baz[2]=b&baz[c[0]]=whatup&baz[c[1]]=notmuch'
-    assert query_string_valid? 'shop=some-random-store.myshopify.com&path_prefix=%2Fapps%2Fmy-app&timestamp=1466106083&foo=bar&baz[1]&baz[2]=b&baz[c[0]]=whatup&baz[c[1]]=notmuch&signature=bbf3aa60e098f08919a2ea4c64a388414f164e6a117a63b03479ac7aa9464b4f'
+    assert query_string_valid?('shop=some-random-store.myshopify.com&path_prefix=%2Fapps%2Fmy-app&timestamp'\
+    '=1466106083&signature=bbf3aa60e098f08919a2ea4c64a388414f164e6a117a63b03479ac7aa9464b4f&foo=bar&baz[1]'\
+    '&baz[2]=b&baz[c[0]]=whatup&baz[c[1]]=notmuch')
+    assert query_string_valid?('shop=some-random-store.myshopify.com&path_prefix=%2Fapps%2Fmy-app&'\
+    'timestamp=1466106083&foo=bar&baz[1]&baz[2]=b&baz[c[0]]=whatup&baz[c[1]]=notmuch&signature'\
+    '=bbf3aa60e098f08919a2ea4c64a388414f164e6a117a63b03479ac7aa9464b4f')
   end
 
   test 'request with invalid signature should fail with 403' do
@@ -41,7 +50,7 @@ class AppProxyVerificationTest < ActionController::TestCase
         shop: 'some-random-store.myshopify.com',
         path_prefix: '/apps/my-app',
         timestamp: '1466106083',
-        signature: 'wrong233558b1c50102a6f33c0b63ad1e1072a2fc126cb58d4500f75223cefcd'
+        signature: 'wrong233558b1c50102a6f33c0b63ad1e1072a2fc126cb58d4500f75223cefcd',
       }
       get :basic, params: invalid_params
       assert_response :forbidden
@@ -54,7 +63,7 @@ class AppProxyVerificationTest < ActionController::TestCase
         shop: 'some-random-store.myshopify.com',
         path_prefix: '/apps/my-app',
         timestamp: '1466106083',
-        signature: 'f5cd7233558b1c50102a6f33c0b63ad1e1072a2fc126cb58d4500f75223cefcd'
+        signature: 'f5cd7233558b1c50102a6f33c0b63ad1e1072a2fc126cb58d4500f75223cefcd',
       }
       get :basic, params: valid_params
       assert_response :ok
@@ -64,7 +73,7 @@ class AppProxyVerificationTest < ActionController::TestCase
   private
 
   def query_string_valid?(query_string)
-    AppProxyVerificationController.new.send(:query_string_valid?,query_string)
+    AppProxyVerificationController.new.send(:query_string_valid?, query_string)
   end
 
   def with_test_routes
