@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'test_helper'
 require 'action_controller'
 require 'action_controller/base'
@@ -6,17 +7,23 @@ require 'action_view/testing/resolvers'
 class LoginProtectionController < ActionController::Base
   include ShopifyApp::EmbeddedApp
   include ShopifyApp::LoginProtection
-  helper_method :shop_session
+  helper_method :current_shopify_session
 
-  around_action :shopify_session, only: [:index]
-  before_action :login_again_if_different_shop, only: [:second_login]
+  around_action :activate_shopify_session, only: [:index]
+  before_action :login_again_if_different_user_or_shop, only: [:second_login]
 
   def index
-    render plain: "OK"
+    render(plain: "OK")
+  end
+
+  def index_with_headers
+    response.set_header('Mock-Header', 'Mock-Value')
+    signal_access_token_required
+    render(plain: "OK")
   end
 
   def second_login
-    render plain: "OK"
+    render(plain: "OK")
   end
 
   def redirect
@@ -24,17 +31,25 @@ class LoginProtectionController < ActionController::Base
   end
 
   def raise_unauthorized
-    raise ActiveResource::UnauthorizedAccess.new('unauthorized')
+    raise ActiveResource::UnauthorizedAccess, 'unauthorized'
   end
 end
 
-class LoginProtectionTest < ActionController::TestCase
+class LoginProtectionControllerTest < ActionController::TestCase
   tests LoginProtectionController
 
   setup do
-    ShopifyApp::SessionRepository.storage = ShopifyApp::InMemorySessionStore
+    ShopifyApp::SessionRepository.shop_storage = ShopifyApp::InMemoryShopSessionStore
+    ShopifyApp::SessionRepository.user_storage = ShopifyApp::InMemoryUserSessionStore
 
-    request.env['HTTP_USER_AGENT'] = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/69.0.3497.100 Safari/537.36'
+    ShopifyApp.configuration.api_key = 'api_key'
+
+    request.env['HTTP_USER_AGENT'] = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_13_6) '\
+                                     'AppleWebKit/537.36 (KHTML, like Gecko) Chrome/69.0.3497.100 Safari/537.36'
+  end
+
+  teardown do
+    ShopifyApp.configuration.allow_jwt_authentication = false
   end
 
   test '#index sets test cookie if embedded app and user agent can partition cookies' do
@@ -54,65 +69,182 @@ class LoginProtectionTest < ActionController::TestCase
     end
   end
 
-  test "#shop_session returns nil when session is nil" do
+  test "#current_shopify_session returns nil when session is nil" do
     with_application_test_routes do
-      session[:shopify] = nil
+      session[:shop_id] = nil
       get :index
-      assert_nil @controller.shop_session
+      assert_nil @controller.current_shopify_session
     end
   end
 
-  test "#shop_session retreives the session from storage" do
+  test "#current_shopify_session retrieves user session using jwt" do
+    ShopifyApp.configuration.allow_jwt_authentication = true
+    domain = 'https://test.myshopify.io'
+    token = 'admin_api_token'
+    dest = 'shopify_domain'
+    sub = 'shopify_user'
+
+    expected_session = ShopifyAPI::Session.new(
+      domain: domain,
+      token: token,
+      api_version: '2020-01',
+    )
+
+    ShopifyApp::SessionRepository.expects(:retrieve_user_session_by_shopify_user_id)
+      .with(sub).returns(expected_session)
+    ShopifyApp::SessionRepository.expects(:retrieve_user_session).never
+    ShopifyApp::SessionRepository.expects(:retrieve_shop_session_by_shopify_domain).never
+    ShopifyApp::SessionRepository.expects(:retrieve_shop_session).never
+
     with_application_test_routes do
-      session[:shopify] = "foobar"
+      request.env['jwt.shopify_domain'] = dest
+      request.env['jwt.shopify_user_id'] = sub
       get :index
-      ShopifyApp::SessionRepository.expects(:retrieve).returns(session).once
-      assert @controller.shop_session
+
+      assert_equal expected_session, @controller.current_shopify_session
     end
   end
 
-  test "#shop_session is memoized and does not retreive session twice" do
+  test "#current_shopify_session retrieves shop session using jwt" do
+    ShopifyApp.configuration.allow_jwt_authentication = true
+    domain = 'https://test.myshopify.io'
+    token = 'admin_api_token'
+    dest = 'test.shopify.com'
+
+    expected_session = ShopifyAPI::Session.new(
+      domain: domain,
+      token: token,
+      api_version: '2020-01',
+    )
+
+    ShopifyApp::SessionRepository.expects(:retrieve_user_session_by_shopify_user_id).never
+    ShopifyApp::SessionRepository.expects(:retrieve_user_session).never
+    ShopifyApp::SessionRepository.expects(:retrieve_shop_session_by_shopify_domain)
+      .with(dest).returns(expected_session)
+    ShopifyApp::SessionRepository.expects(:retrieve_shop_session).never
+
     with_application_test_routes do
-      session[:shopify] = "foobar"
+      request.env['jwt.shopify_domain'] = dest
       get :index
-      ShopifyApp::SessionRepository.expects(:retrieve).returns(session).once
-      assert @controller.shop_session
-      assert @controller.shop_session
+
+      assert_equal expected_session, @controller.current_shopify_session
     end
   end
 
-  test "#login_again_if_different_shop removes current session and redirects to login url" do
+  test "#current_shopify_session retrieves using session user_id" do
     with_application_test_routes do
-      session[:shopify] = "foobar"
+      session[:user_id] = '145'
+      get :index
+      ShopifyApp::SessionRepository.expects(:retrieve_user_session).with(session[:user_id]).returns(session).once
+      assert @controller.current_shopify_session
+    end
+  end
+
+  test "#current_shopify_session retrieves using shop_id when shopify_user not present" do
+    with_application_test_routes do
+      session[:shop_id] = "shopify_id"
+      get :index
+      ShopifyApp::SessionRepository.expects(:retrieve_shop_session).with(session[:shop_id]).returns(session).once
+      assert @controller.current_shopify_session
+    end
+  end
+
+  test "#current_shopify_session retreives the session from storage" do
+    with_application_test_routes do
+      session[:shop_id] = "foobar"
+      get :index
+      ShopifyApp::SessionRepository.expects(:retrieve_shop_session).returns(session).once
+      assert @controller.current_shopify_session
+    end
+  end
+
+  test "#current_shopify_session is memoized and does not retreive session twice" do
+    with_application_test_routes do
+      session[:shop_id] = "foobar"
+      get :index
+      ShopifyApp::SessionRepository.expects(:retrieve_shop_session).returns(session).once
+      assert @controller.current_shopify_session
+    end
+  end
+
+  test "#login_again_if_different_user_or_shop removes current session if the user changes when in per-user-token mode" do
+    with_application_test_routes do
+      session[:shop_id] = "1"
+      session[:shopify_domain] = "foobar"
+      session[:shopify_user] = { 'id' => 1, 'email' => 'foo@example.com' }
+      session[:user_session] = 'old-user-session'
+      params = { shop: 'foobar', session: 'new-user-session' }
+      get :second_login, params: params
+      assert_nil session[:shop_id]
+      assert_nil session[:shopify_domain]
+      assert_nil session[:shopify_user]
+      assert_nil session[:user_session]
+    end
+  end
+
+  test "#login_again_if_different_user_or_shop retains current session if the users session doesn't change" do
+    with_application_test_routes do
+      session[:shop_id] = "1"
+      session[:shopify_domain] = "foobar"
+      session[:shopify_user] = { 'id' => 1, 'email' => 'foo@example.com' }
+      session[:user_session] = 'old-user-session'
+      params = { shop: 'foobar', session: 'old-user-session' }
+      get :second_login, params: params
+      assert session[:shop_id], "1"
+      assert session[:shopify_domain], "foobar"
+      assert session[:shopify_user], { 'id' => 1, 'email' => 'foo@example.com' }
+      assert session[:user_session], 'old-user-session'
+    end
+  end
+
+  test "#login_again_if_different_user_or_shop retains current session if params not present" do
+    with_application_test_routes do
+      session[:shop_id] = "1"
+      session[:shopify_domain] = "foobar"
+      session[:shopify_user] = { 'id' => 1, 'email' => 'foo@example.com' }
+      session[:user_session] = 'old-user-session'
+      get :second_login
+      assert session[:shop_id], "1"
+      assert session[:shopify_domain], "foobar"
+      assert session[:shopify_user], { 'id' => 1, 'email' => 'foo@example.com' }
+      assert session[:user_session], 'old-user-session'
+    end
+  end
+
+  test "#login_again_if_different_user_or_shop removes current session and redirects to login url" do
+    with_application_test_routes do
+      session[:shop_id] = "foobar"
+      session[:user_id] = 123
       session[:shopify_domain] = "foobar"
       session[:shopify_user] = { 'id' => 1, 'email' => 'foo@example.com' }
       sess = stub(domain: 'https://foobar.myshopify.com')
-      ShopifyApp::SessionRepository.expects(:retrieve).returns(sess).once
+      ShopifyApp::SessionRepository.expects(:retrieve_user_session).returns(sess).once
       get :second_login, params: { shop: 'other-shop' }
-      assert_redirected_to '/login?return_to=%2Fsecond_login%3Fshop%3Dother-shop.myshopify.com&shop=other-shop.myshopify.com'
-      assert_nil session[:shopify]
+      assert_redirected_to '/login?return_to=%2Fsecond_login%3Fshop%3Dother-shop.myshopify.com'\
+                           '&shop=other-shop.myshopify.com'
+      assert_nil session[:shop_id]
       assert_nil session[:shopify_domain]
       assert_nil session[:shopify_user]
     end
   end
 
-  test "#login_again_if_different_shop ignores non-String shop params so that Rails params for Shop model can be accepted" do
+  test "#login_again_if_different_user_or_shop ignores non-String shop params so that Rails params for Shop model can be accepted" do
     with_application_test_routes do
-      session[:shopify] = "foobar"
+      session[:shop_id] = "foobar"
       session[:shopify_domain] = "foobar"
       sess = stub(domain: 'https://foobar.myshopify.com')
-      ShopifyApp::SessionRepository.expects(:retrieve).returns(sess).once
+      ShopifyApp::SessionRepository.expects(:retrieve_shop_session).returns(sess).once
 
       get :second_login, params: { shop: { id: 123, disabled: true } }
       assert_response :ok
     end
   end
 
-  test '#shopify_session with Shopify session, clears top-level auth cookie' do
+  test '#activate_shopify_session with Shopify session, clears top-level auth cookie' do
     with_application_test_routes do
       session['shopify.top_level_oauth'] = true
       sess = stub(domain: 'https://foobar.myshopify.com')
-      @controller.expects(:shop_session).returns(sess).at_least_once
+      @controller.expects(:current_shopify_session).returns(sess).at_least_once
       ShopifyAPI::Base.expects(:activate_session).with(sess)
 
       get :index, params: { shop: 'foobar' }
@@ -120,14 +252,14 @@ class LoginProtectionTest < ActionController::TestCase
     end
   end
 
-  test '#shopify_session with no Shopify session, redirects to the login url' do
+  test '#activate_shopify_session with no Shopify session, redirects to the login url' do
     with_application_test_routes do
       get :index, params: { shop: 'foobar' }
       assert_redirected_to '/login?shop=foobar.myshopify.com'
     end
   end
 
-  test '#shopify_session with no Shopify session, redirects to a custom config login url' do
+  test '#activate_shopify_session with no Shopify session, redirects to a custom config login url' do
     with_custom_login_url 'https://domain.com/custom/route/login' do
       with_application_test_routes do
         get :index, params: { shop: 'foobar' }
@@ -136,10 +268,10 @@ class LoginProtectionTest < ActionController::TestCase
     end
   end
 
-  test "#shopify_session with no Shopify session, redirects to login_url with \
+  test "#activate_shopify_session with no Shopify session, redirects to login_url with \
         shop param of referer" do
     with_application_test_routes do
-      @controller.expects(:shop_session).returns(nil)
+      @controller.expects(:current_shopify_session).returns(nil)
       request.headers['Referer'] = 'https://example.com/?shop=my-shop.myshopify.com'
 
       get :index
@@ -147,11 +279,11 @@ class LoginProtectionTest < ActionController::TestCase
     end
   end
 
-  test "#shopify_session with no Shopify session, redirects to a custom config login url with \
+  test "#activate_shopify_session with no Shopify session, redirects to a custom config login url with \
         shop param of referer" do
     with_custom_login_url 'https://domain.com/custom/route/login' do
       with_application_test_routes do
-        @controller.expects(:shop_session).returns(nil)
+        @controller.expects(:current_shopify_session).returns(nil)
         request.headers['Referer'] = 'https://example.com/?shop=my-shop.myshopify.com'
 
         get :index
@@ -160,7 +292,7 @@ class LoginProtectionTest < ActionController::TestCase
     end
   end
 
-  test '#shopify_session with no Shopify session, redirects to the login url \
+  test '#activate_shopify_session with no Shopify session, redirects to the login url \
         with non-String shop param' do
     with_application_test_routes do
       params = { shop: { id: 123 } }
@@ -169,7 +301,7 @@ class LoginProtectionTest < ActionController::TestCase
     end
   end
 
-  test '#shopify_session with no Shopify session, redirects to a custom config login url \
+  test '#activate_shopify_session with no Shopify session, redirects to a custom config login url \
         with non-String shop param' do
     with_custom_login_url 'https://domain.com/custom/route/login' do
       with_application_test_routes do
@@ -180,14 +312,14 @@ class LoginProtectionTest < ActionController::TestCase
     end
   end
 
-  test '#shopify_session with no Shopify session, sets session[:return_to]' do
+  test '#activate_shopify_session with no Shopify session, sets session[:return_to]' do
     with_application_test_routes do
       get :index, params: { shop: 'foobar' }
       assert_equal '/?shop=foobar.myshopify.com', session[:return_to]
     end
   end
 
-  test '#shopify_session with no Shopify session, sets session[:return_to]\
+  test '#activate_shopify_session with no Shopify session, sets session[:return_to]\
         with non-String shop param' do
     with_application_test_routes do
       params = { shop: { id: 123 } }
@@ -196,29 +328,37 @@ class LoginProtectionTest < ActionController::TestCase
     end
   end
 
-  test '#shopify_session with no Shopify session, when the request is an XHR, returns an HTTP 401' do
+  test '#activate_shopify_session with no Shopify session, when the request is a POST, sets session[:return_to]' do
+    with_application_test_routes do
+      request.headers['Referer'] = 'https://example.com/?id=123'
+      post :index, params: { id: '123', shop: 'foobar' }
+      assert_equal '/?id=123&shop=foobar.myshopify.com', session[:return_to]
+    end
+  end
+
+  test '#activate_shopify_session with no Shopify session, when the request is an XHR, returns an HTTP 401' do
     with_application_test_routes do
       get :index, params: { shop: 'foobar' }, xhr: true
       assert_equal 401, response.status
     end
   end
 
-  test '#shopify_session when rescuing from unauthorized access, redirects to the login url' do
+  test '#activate_shopify_session when rescuing from unauthorized access, redirects to the login url' do
     with_application_test_routes do
       get :raise_unauthorized, params: { shop: 'foobar' }
       assert_redirected_to '/login?shop=foobar.myshopify.com'
     end
   end
 
-  test '#shopify_session when rescuing from unauthorized access, clears shop session' do
+  test '#activate_shopify_session when rescuing from unauthorized access, clears shop session' do
     with_application_test_routes do
-      session[:shopify] = 'foobar'
+      session[:shop_id] = 'foobar'
       session[:shopify_domain] = 'foobar'
       session[:shopify_user] = { 'id' => 1, 'email' => 'foo@example.com' }
 
       get :raise_unauthorized, params: { shop: 'foobar' }
 
-      assert_nil session[:shopify]
+      assert_nil session[:shop_id]
       assert_nil session[:shopify_domain]
       assert_nil session[:shopify_user]
     end
@@ -229,6 +369,17 @@ class LoginProtectionTest < ActionController::TestCase
       example_shop = 'shop.myshopify.com'
       get :redirect, params: { shop: example_shop }
       assert_fullpage_redirected(example_shop, response)
+    end
+  end
+
+  test '#fullpage_redirect_to, when the shop params is missing, sends a post message to the shop in the jwt' do
+    ShopifyApp.configuration.allow_jwt_authentication = true
+    domain = 'shop.myshopify.com'
+
+    with_application_test_routes do
+      request.env['jwt.shopify_domain'] = domain
+      get :redirect
+      assert_fullpage_redirected(domain, response)
     end
   end
 
@@ -268,12 +419,26 @@ class LoginProtectionTest < ActionController::TestCase
     end
   end
 
+  test 'signal_access_token_required sets X-Shopify-API-Request-Unauthorized header' do
+    with_application_test_routes do
+      get :index_with_headers
+      assert_equal true, response.get_header('X-Shopify-API-Request-Failure-Unauthorized')
+    end
+  end
+
+  test 'signal_access_token_required does not overwrite previously set headers' do
+    with_application_test_routes do
+      get :index_with_headers
+      assert_equal 'Mock-Value', response.get_header('Mock-Header')
+    end
+  end
+
   private
 
-  def assert_fullpage_redirected(shop_domain, response)
+  def assert_fullpage_redirected(shop_domain, _response)
     example_url = "https://example.com"
 
-    assert_template 'shared/redirect'
+    assert_template('shared/redirect')
     assert_select '[id=redirection-target]', 1 do |elements|
       assert_equal "{\"myshopifyUrl\":\"https://#{shop_domain}\",\"url\":\"#{example_url}\"}",
         elements.first['data-target']
@@ -287,6 +452,7 @@ class LoginProtectionTest < ActionController::TestCase
         get '/second_login' => 'login_protection#second_login'
         get '/redirect' => 'login_protection#redirect'
         get '/raise_unauthorized' => 'login_protection#raise_unauthorized'
+        get '/index_with_headers' => 'login_protection#index_with_headers'
       end
       yield
     end
