@@ -1,12 +1,12 @@
+# frozen_string_literal: true
 require 'test_helper'
 
 class ShopifyApp::WebhooksManagerTest < ActiveSupport::TestCase
-
   setup do
     @webhooks = [
-      {topic: 'app/uninstalled', address: "https://example-app.com/webhooks/app_uninstalled"},
-      {topic: 'orders/create', address: "https://example-app.com/webhooks/order_create"},
-      {topic: 'orders/updated', address: "https://example-app.com/webhooks/order_updated"},
+      { topic: 'app/uninstalled', address: "https://example-app.com/webhooks/app_uninstalled" },
+      { topic: 'orders/create', address: "https://example-app.com/webhooks/order_create" },
+      { topic: 'orders/updated', address: "https://example-app.com/webhooks/order_updated" },
     ]
 
     @manager = ShopifyApp::WebhooksManager.new(@webhooks)
@@ -72,7 +72,8 @@ class ShopifyApp::WebhooksManagerTest < ActiveSupport::TestCase
   end
 
   test "#destroy_webhooks does not destroy webhooks that do not have a matching address" do
-    ShopifyAPI::Webhook.stubs(:all).returns([stub(address: 'http://something-or-the-other.com/webhooks/product_update', id: 7214109)])
+    ShopifyAPI::Webhook.stubs(:all).returns([stub(address: 'http://something-or-the-other.com/webhooks/product_update',
+                                                  id: 7214109)])
     ShopifyAPI::Webhook.expects(:delete).never
 
     @manager.destroy_webhooks
