@@ -1,12 +1,14 @@
+# frozen_string_literal: true
 require 'test_helper'
 
 class UtilsTest < ActiveSupport::TestCase
-
   setup do
     ShopifyApp.configuration = nil
   end
 
-  ['my-shop', 'my-shop.myshopify.com', 'https://my-shop.myshopify.com', 'http://my-shop.myshopify.com'].each do |good_url|
+  ['my-shop', 'my-shop.myshopify.com',
+   'https://my-shop.myshopify.com',
+   'http://my-shop.myshopify.com'].each do |good_url|
     test "sanitize_shop_domain for (#{good_url})" do
       assert ShopifyApp::Utils.sanitize_shop_domain(good_url)
     end
@@ -23,7 +25,9 @@ class UtilsTest < ActiveSupport::TestCase
     assert ShopifyApp::Utils.sanitize_shop_domain('MY-shop.myshopify.com')
   end
 
-  ['myshop.com', 'myshopify.com', 'shopify.com', 'two words', 'store.myshopify.com.evil.com', '/foo/bar', 'foo.myshopify.io.evil.ru', '%0a123.myshopify.io', 'foo.bar.myshopify.io'].each do |bad_url|
+  ['myshop.com', 'myshopify.com', 'shopify.com', 'two words', 'store.myshopify.com.evil.com',
+   '/foo/bar', 'foo.myshopify.io.evil.ru', '%0a123.myshopify.io',
+   'foo.bar.myshopify.io'].each do |bad_url|
     test "sanitize_shop_domain for a non-myshopify URL (#{bad_url})" do
       assert_nil ShopifyApp::Utils.sanitize_shop_domain(bad_url)
     end
