@@ -1,9 +1,10 @@
+# frozen_string_literal: true
 # Configure Rails Environment
 ENV["RAILS_ENV"] = "test"
 
-require File.expand_path("../../test/dummy/config/environment.rb",  __FILE__)
+require File.expand_path("../../test/dummy/config/environment.rb", __FILE__)
 require 'rails/test_help'
-require 'mocha/setup'
+require 'mocha/minitest'
 require 'webmock/minitest'
 require 'byebug'
 require 'pry'
@@ -12,30 +13,30 @@ require 'pry'
 # to be shown.
 Minitest.backtrace_filter = Minitest::BacktraceFilter.new
 
-Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each {|f| require f}
+Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each { |f| require f }
 
 class ActiveSupport::TestCase
   include GeneratorTestHelpers
+  include SessionStoreStrategyTestHelpers
 
-  API_META_TEST_RESPONSE =  <<~JSON
-  {
-    "apis": [{
-      "handle": "admin",
-      "versions": [{
-        "handle": "2019-07",
-        "display_name": "2019-07",
-        "supported": true,
-        "latest_supported": false
-      },{
-          "handle": "2019-10",
-          "latest_supported": true,
-          "display_name": "2019-10",
-          "supported": true
+  API_META_TEST_RESPONSE = <<~JSON
+    {
+      "apis": [{
+        "handle": "admin",
+        "versions": [{
+          "handle": "2019-07",
+          "display_name": "2019-07",
+          "supported": true,
+          "latest_supported": false
+        },{
+            "handle": "2019-10",
+            "latest_supported": true,
+            "display_name": "2019-10",
+            "supported": true
+          }]
         }]
-      }]
-    }
+      }
   JSON
-
 
   def before_setup
     super
